Abstracted code to its own file
