+++ src/helpers/Sum.js
@@ -1,0 +1,7 @@
+const Sum = list => {
+  return list.reduce(function(prev, cur, index, array) {
+    return prev + cur;
+  });
+};
+
+export default Sum;
